refactor(tournaments): split tournaments.py (2,417 lines) into 10 domain modules
Decomposes the monolithic web_routes/tournaments.py into a structured package:

  app/api/web_routes/tournaments/
  ├── __init__.py            — shared helpers + sub-module aggregator
  ├── browse.py              — student tournament browse/enroll (3 routes)
  ├── camps.py               — student camp browse/enroll (3 routes)
  ├── instructor.py          — instructor tournaments view (1 route)
  ├── lifecycle.py           — admin CRUD + status transitions (7 routes)
  ├── edit.py                — admin tournament edit page (1 route)
  ├── instructors_admin.py   — admin instructor list + detail (2 routes)
  ├── teams.py               — admin team enrollment management (7 routes)
  ├── players.py             — admin individual player enrollment (8 routes)
  ├── instructor_planning.py — instructor slot management IP-* (7 routes)
  └── attendance.py          — attendance + check-in endpoints (6 routes)

Validation gates:
  ✅ 1,737/1,737 API smoke tests
  ✅ OpenAPI zero-diff (661 routes)
  ✅ 9/9 query budget tests
  Zero behavior change — all 45 routes intact
