+++ app/api/web_routes/__init__.py
@@ -16,7 +16,6 @@ from . import (
     quiz,
     instructor,
     instructor_dashboard,
-    tournaments,
     programs,
     communications,
     teams,
@@ -26,6 +25,7 @@ from . import (
     sport_director,
 )
 from .admin import router as admin_router
+from .tournaments import router as tournaments_router
 
 # Create main router with tags
 router = APIRouter(tags=["web"])
@@ -43,7 +43,7 @@ router.include_router(quiz.router)
 router.include_router(instructor.router)
 router.include_router(instructor_dashboard.router)
 router.include_router(admin_router)
-router.include_router(tournaments.router)
+router.include_router(tournaments_router)
 router.include_router(programs.router)       # 📅 MINI_SEASON / ACADEMY_SEASON student enrollment
 router.include_router(communications.router)
 router.include_router(teams.router)
